+++ News.md
@@ -11,5 +11,5 @@ Initial released version is [`1.0.2`](https://pypi.org/project/labw-utils/1.0.2/
 - `1.0.3`:
   - Documentation enhancements for `labw_utils.PackageSpecs`.
   - A command-line interface added to `labw_utils.commonutils`
-  - Enum support added to `labw_utils.commonutils.stdlib_helper.argparse_helper`.
+  - Enum support added to `labw_utils.commonutils.stdlib_helper.argparse_helper`; Setting both `required` and `default` would generate `ValueError`.
   - Miscellaneous bug fixes.